feat: add registration screen
