+++ app/src/main/java/it/android/servarium/navigation/Route.kt
@@ -1,5 +1,6 @@
 package it.android.servarium.navigation
 
 sealed class Route {
-    data class AuthScreen(val route: String = "auth") : Route()
+    data class LoginScreen(val route: String = "login") : Route()
+    data class RegistrationScreen(val route: String = "registration") : Route()
 }
